refactor: Made Requested Changes
